+++ src/ralph/discovery/tests/__init__.py
@@ -1,5 +1,1 @@
 # -*- coding: utf-8 -*-
-
-from ralph.discovery.tests.plugins import *
-from ralph.discovery.tests.model_tests import ModelsTest
-from ralph.discovery.tests.dmidecode_tests import DMIDecodeTest